Actu

+++ Pocion.py
@@ -2,7 +2,7 @@ import random
 
 hp = 50
 dificultad = 3
-pocion_salud = random.randint(25,50) / dificultad
+pocion_salud = int(random.randint(25,50) / dificultad)
 
 hp += pocion_salud
 print (hp)
